requirements.txt
django
uvicorn
djangorestframework
django-rest-framework-mongoengine
mongoengine
httptools
pydantic-settings
django-cors-headers